tetragon: Move Enable* functions to config package
Moving Enable* functions to config package, because it makes
sense and we get bpf package dependency out of kernels package.
