+++ pkg/selectors/kernel.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/cilium/tetragon/api/v1/tetragon"
 	"github.com/cilium/tetragon/pkg/api/processapi"
+	"github.com/cilium/tetragon/pkg/config"
 	gt "github.com/cilium/tetragon/pkg/generictypes"
 	"github.com/cilium/tetragon/pkg/idtable"
 	"github.com/cilium/tetragon/pkg/k8s/apis/cilium.io/v1alpha1"
@@ -757,7 +758,7 @@ func writePostfixStrings(k *KernelSelectorState, values []string, ty uint32) err
 func checkOp(op uint32) error {
 	switch op {
 	case SelectorOpGT, SelectorOpLT:
-		if !kernels.EnableLargeProgs() {
+		if !config.EnableLargeProgs() {
 			return fmt.Errorf("GT/LT operators are only supported in kernels >= 5.3")
 		}
 	}
@@ -854,7 +855,7 @@ func ParseMatchArg(k *KernelSelectorState, arg *v1alpha1.ArgSelector, sig []v1al
 
 func ParseMatchArgs(k *KernelSelectorState, args []v1alpha1.ArgSelector, sig []v1alpha1.KProbeArg) error {
 	max_args := 1
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		max_args = 5 // we support up 5 argument filters under matchArgs with kernels >= 5.3, otherwise 1 argument
 	}
 	if len(args) > max_args {
@@ -1077,7 +1078,7 @@ func ParseMatchNamespace(k *KernelSelectorState, action *v1alpha1.NamespaceSelec
 
 func ParseMatchNamespaces(k *KernelSelectorState, actions []v1alpha1.NamespaceSelector) error {
 	max_nactions := 4 // 4 should match the value of the NUM_NS_FILTERS_SMALL in pfilter.h
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		max_nactions = 10 // 10 should match the value of ns_max_types in hubble_msg.h
 	}
 	if len(actions) > max_nactions {
@@ -1123,7 +1124,7 @@ func ParseMatchNamespaceChanges(k *KernelSelectorState, actions []v1alpha1.Names
 	if len(actions) > 1 {
 		return fmt.Errorf("matchNamespaceChanges supports only a single filter (current number of filters is %d)", len(actions))
 	}
-	if (len(actions) == 1) && !kernels.EnableLargeProgs() {
+	if (len(actions) == 1) && !config.EnableLargeProgs() {
 		return fmt.Errorf("matchNamespaceChanges is only supported in kernels >= 5.3")
 	}
 	loff := AdvanceSelectorLength(&k.data)
@@ -1238,7 +1239,7 @@ func ParseMatchBinary(k *KernelSelectorState, b *v1alpha1.BinarySelector, selIdx
 			k.WriteMatchBinariesPath(selIdx, s)
 		}
 	case SelectorOpPrefix, SelectorOpNotPrefix:
-		if !kernels.EnableLargeProgs() {
+		if !config.EnableLargeProgs() {
 			return fmt.Errorf("matchBinary error: \"Prefix\" and \"NotPrefix\" operators need large BPF progs (kernel>5.3)")
 		}
 		sel.MapID, err = writePrefixBinaries(k, b.Values)
@@ -1246,7 +1247,7 @@ func ParseMatchBinary(k *KernelSelectorState, b *v1alpha1.BinarySelector, selIdx
 			return fmt.Errorf("failed to write the prefix operator for the matchBinaries selector: %w", err)
 		}
 	case SelectorOpPostfix, SelectorOpNotPostfix:
-		if !kernels.EnableLargeProgs() {
+		if !config.EnableLargeProgs() {
 			return fmt.Errorf("matchBinary error: \"Postfix\" and \"NotPostfix\" operators need large BPF progs (kernel>5.3)")
 		}
 		sel.MapID, err = writePostfixBinaries(k, b.Values)
