+++ pkg/selectors/kernel_test.go
@@ -14,10 +14,10 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/cilium/tetragon/pkg/config"
 	gt "github.com/cilium/tetragon/pkg/generictypes"
 	"github.com/cilium/tetragon/pkg/idtable"
 	"github.com/cilium/tetragon/pkg/k8s/apis/cilium.io/v1alpha1"
-	"github.com/cilium/tetragon/pkg/kernels"
 )
 
 func TestWriteSelectorUint32(t *testing.T) {
@@ -350,7 +350,7 @@ func TestParseMatchArg(t *testing.T) {
 		t.Errorf("parseMatchArg: error %v expected %v bytes %v parsing %v\n", err, expected8, d.e[nextArg:d.off], arg8)
 	}
 
-	if kernels.EnableLargeProgs() { // multiple match args are supported only in kernels >= 5.4
+	if config.EnableLargeProgs() { // multiple match args are supported only in kernels >= 5.4
 		length := []byte{
 			108, 0x00, 0x00, 0x00,
 			24, 0x00, 0x00, 0x00,
@@ -811,7 +811,7 @@ func TestInitKernelSelectors(t *testing.T) {
 	}
 
 	expected := expected_header
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		expected = append(expected, expected_selsize_large...)
 		expected = append(expected, expected_filters...)
 		expected = append(expected, expected_changes...)
@@ -833,17 +833,17 @@ func TestInitKernelSelectors(t *testing.T) {
 	cap2 := &v1alpha1.CapabilitiesSelector{Type: "Inheritable", Operator: "NotIn", IsNamespaceCapability: false, Values: []string{"CAP_SETPCAP", "CAP_SYS_ADMIN"}}
 	matchCapabilities := []v1alpha1.CapabilitiesSelector{*cap1, *cap2}
 	matchNamespaceChanges := []v1alpha1.NamespaceChangesSelector{}
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		nc := &v1alpha1.NamespaceChangesSelector{Operator: "In", Values: []string{"Uts", "Mnt"}}
 		matchNamespaceChanges = append(matchNamespaceChanges, *nc)
 	}
 	matchCapabilityChanges := []v1alpha1.CapabilitiesSelector{}
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		cc := &v1alpha1.CapabilitiesSelector{Type: "Effective", Operator: "In", IsNamespaceCapability: false, Values: []string{"CAP_SYS_ADMIN", "CAP_NET_RAW"}}
 		matchCapabilityChanges = append(matchCapabilityChanges, *cc)
 	}
 	var matchArgs []v1alpha1.ArgSelector
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		arg1 := &v1alpha1.ArgSelector{Index: 1, Operator: "Equal", Values: []string{"foobar"}}
 		arg2 := &v1alpha1.ArgSelector{Index: 2, Operator: "Equal", Values: []string{"1", "2"}}
 		matchArgs = []v1alpha1.ArgSelector{*arg1, *arg2}
