+++ pkg/sensors/tracing/kprobe_validation_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/cilium/tetragon/pkg/bpf"
-	"github.com/cilium/tetragon/pkg/kernels"
+	"github.com/cilium/tetragon/pkg/config"
 	"github.com/cilium/tetragon/pkg/sensors"
 	"github.com/cilium/tetragon/pkg/tracingpolicy"
 	"github.com/stretchr/testify/assert"
@@ -279,7 +279,7 @@ spec:
 `
 
 	err := checkCrd(t, crd)
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		assert.NoError(t, err)
 	} else {
 		assert.Error(t, err)
@@ -310,7 +310,7 @@ spec:
 `
 
 	err := checkCrd(t, crd)
-	if kernels.EnableLargeProgs() {
+	if config.EnableLargeProgs() {
 		assert.NoError(t, err)
 	} else {
 		assert.Error(t, err)
